requirements.txt
django-environ==0.4.5
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
